ci: Conditionally build parallel compiler on `try`
This commit configures Travis/AppVeyor to conditionally compile parallel
compilers on `@bors: try`. This is an experiment currently to see how
this plays out, but the intention is that if the commit message contains
the term "parallel-compiler" then when `@bors: try` is issued it will
perform differently than the try branch does today, building three
compilers: Linux, macOS, and Windows. We currently have no `try`
builders for macOS or Windows due to typical capacity issues, so it's
intended that this is only very sparingly used from time to time when
necessary.

[parallel-compiler]